ktl-4625 feat: rename kotlin-benchmark to benchmark
