+++ blocks/benchmark/methodology/hero/index.tsx
